Backbone (not prototype port) with old sync method. Read notes
Ok, so this is a little bit messy, but anyway, we'll have to upgrade
to 0.5 which should be compatible with this changes. The thing is that
with the backbone-prototype's Backbone.sync, an url could be set as an
option. In plain backbone 0.3.2 this can not be done. With this
changes we can set an url again. In backbone 0.5.3 (latest version atm)
Backbone.sync works this way, so we should not have this problem
when we upgrade. Happy coding! :)